Fix big digits comma formatting for popups and labels
